Department_relation_type_control done vali

+++ AdmissionSystem/AdmissionSystem/Controllers/sub_classes/Department_relation_type_control.cs
@@ -1,4 +1,5 @@
-﻿using AdmissionSystem.Model;
+﻿using AdmissionSystem.Data;
+using AdmissionSystem.Model;
 using AdmissionSystem.Model.Repository;
 using AdmissionSystem.View_Model;
 using Microsoft.AspNetCore.Http;
@@ -15,15 +16,18 @@ namespace AdmissionSystem.Controllers.sub_classes.Admin_classes
         private readonly CRUD_Operation_Interface<Department_relation_Type> depart_Relat_Ty_repo;
         private readonly CRUD_Operation_Interface<Department> department_Repo;
         private readonly CRUD_Operation_Interface<Type_of_high_school_Cirtificate> type_Of_High_School_Repo;
+        private readonly ApplicationDbContext dB;
 
         public Department_relation_type_control(CRUD_Operation_Interface<Department_relation_Type> depart_relat_Ty_repo
             ,CRUD_Operation_Interface<Department> department_repo
             ,CRUD_Operation_Interface<Type_of_high_school_Cirtificate> Type_of_high_school_repo
+            , ApplicationDbContext _DB
             )
         {
             depart_Relat_Ty_repo = depart_relat_Ty_repo;
             department_Repo = department_repo;
             type_Of_High_School_Repo = Type_of_high_school_repo;
+            dB = _DB;
         }
         // GET: Department_relation_type_control
         public ActionResult Index()
@@ -69,8 +73,37 @@ namespace AdmissionSystem.Controllers.sub_classes.Admin_classes
                 Minemum_of_Rate=collection.Minemum_of_Rate,
                 Rate_of_chaire_count=collection.Rate_of_chaire_count
                 };
+                var ddrrtt = depart_Relat_Ty_repo.List().Where(d=>d.FK_DepartmentId==collection.Department_id
+                                                               && d.FK_type_Of_High_School_CirtificateId==collection.Type_of_high_school_cirtificate_id
+                                                               && d.Minemum_of_Rate==collection.Minemum_of_Rate
+                                                               && d.Rate_of_chaire_count==collection.Rate_of_chaire_count
+                                                                                ).ToList();
+
+                if (ddrrtt.Count == 0)
+                {
                 depart_Relat_Ty_repo.Add(depar_type_rel);
-                return RedirectToAction(nameof(Index));
+                    ViewBag.AddSuccess = "The addition succeeded";
+                    ViewBag.AddSuccessArabic = "نجحت الإضافة";
+
+                    return View(GetAll());
+                }
+                else if (ddrrtt.Count == 1)
+                {
+                    ViewBag.sameEdit = "Already Exists";
+                    ViewBag.sameEditArabic = "موجود بالفعل";
+
+                    return View(GetAll());
+
+                }
+                else
+                {
+                    ViewBag.Addfails = "The addition fails";
+                    ViewBag.AddfailsArabic = "فشل الإضافة";
+
+                    return View(GetAll());
+
+                }
+                //return RedirectToAction(nameof(Index));
             }
             catch
             {
@@ -82,9 +115,11 @@ namespace AdmissionSystem.Controllers.sub_classes.Admin_classes
         public ActionResult Edit(int id)
         {
             var deprtfind = depart_Relat_Ty_repo.Find(id);
+            var dep = department_Repo.Find(deprtfind.FK_DepartmentId);
+            var type = type_Of_High_School_Repo.Find(deprtfind.FK_type_Of_High_School_CirtificateId);
             var depview = new Department_relation_type_view_model { 
-            FK_Department=FillSellectionDepartment(),
-            FK_type_Of_High_School_Cirtificate=FillSellectionType_of_high_school___(),
+            FK_Department=FillSellectionDepartmentEDit(dep),
+            FK_type_Of_High_School_Cirtificate=FillSellectionType_of_high_school___EDIt(type),
             Minemum_of_Rate=deprtfind.Minemum_of_Rate,
           Rate_of_chaire_count=deprtfind.Rate_of_chaire_count,   
             
@@ -108,10 +143,47 @@ namespace AdmissionSystem.Controllers.sub_classes.Admin_classes
                 Minemum_of_Rate=collection.Minemum_of_Rate,
                 Rate_of_chaire_count=collection.Rate_of_chaire_count
                 };
+                var ddrrtt = depart_Relat_Ty_repo.List().Where(d => d.FK_DepartmentId == collection.Department_id
+                                                              && d.FK_type_Of_High_School_CirtificateId == collection.Type_of_high_school_cirtificate_id
+                                                              && d.Minemum_of_Rate == collection.Minemum_of_Rate
+                                                              && d.Rate_of_chaire_count == collection.Rate_of_chaire_count
+                                                                               ).ToList();
+
+
+                if (ddrrtt.Count == 0)
+                {
+
+
                 depart_Relat_Ty_repo.Update(id,dep);
-                return RedirectToAction(nameof(Index));
+                   
+                    ViewBag.UpdateSuccess = "Editing success";
+                    ViewBag.UpdateSuccessArabic = "نجاح التحرير";
+
+                    return View(GetAllEDit(departfind, typfind));
+                }
+                else if (ddrrtt.Count == 1)
+                {
+                    ViewBag.sameEdit = "Same Edit";
+                    ViewBag.sameEditArabic = "نفس التحرير";
+
+                    return View(GetAllEDit(departfind, typfind));
+
+                }
+                else
+                {
+
+                    ViewBag.updatefails = "Editing failed";
+                    ViewBag.updatefailsArabic = "فشل التحرير";
+
+                    return View(GetAllEDit(departfind, typfind));
+
+                }
+
+
+
+                //return RedirectToAction(nameof(Index));
             }
-            catch
+            catch(Exception e)
             {
                 return View();
             }
@@ -131,8 +203,24 @@ namespace AdmissionSystem.Controllers.sub_classes.Admin_classes
         {
             try
             {
-                depart_Relat_Ty_repo.Delete(id);
-                return RedirectToAction(nameof(Index));
+                var depfind = depart_Relat_Ty_repo.Find(id);
+                if (depfind == null)
+                {
+                    ViewBag.WasDeleted = "This item is already deleted";
+                    ViewBag.WasDeletedArabic = "تم حذف هذا العنصر بالفعل";
+
+                    return View();
+                }
+                else
+                {
+              
+                    depart_Relat_Ty_repo.Delete(id);
+                    ViewBag.freeDepartment = "Delete the item succeeded";
+                    ViewBag.freeDepartmentArabic = "تم الحذف";
+                    return View();
+
+                }
+                //return RedirectToAction(nameof(Index));
             }
             catch
             {
@@ -144,11 +232,29 @@ namespace AdmissionSystem.Controllers.sub_classes.Admin_classes
             deprt.Insert(0,new Department { id =-1,specialization_name= "_-_-_-_-please Enter specialization------" });
             return deprt;
         }
+        List<Department> FillSellectionDepartmentEDit(Department d)
+        {
+            var deprt = dB.Department.ToList();
+            deprt.Remove(d);
+            deprt.Insert(0,d);
+
+            //deprt.Insert(0, new Department { id = -1, specialization_name = "_-_-_-_-please Enter specialization------" });
+            return deprt;
+        }
         List<Type_of_high_school_Cirtificate> FillSellectionType_of_high_school___() {
             var type = type_Of_High_School_Repo.List().ToList();
             type.Insert(0,new Type_of_high_school_Cirtificate {id=-1,type= "_-_-_-_-please Enter the type of hight schoole----" });
             return type;
         
+        }
+        List<Type_of_high_school_Cirtificate> FillSellectionType_of_high_school___EDIt(Type_of_high_school_Cirtificate t)
+        {
+            var type = type_Of_High_School_Repo.List().ToList();
+            type.Remove(t);
+            type.Insert(0,t);
+            //type.Insert(0, new Type_of_high_school_Cirtificate { id = -1, type = "_-_-_-_-please Enter the type of hight schoole----" });
+            return type;
+
         }
         Department_relation_type_view_model GetAll() {
             var dep = new Department_relation_type_view_model
@@ -156,6 +262,16 @@ namespace AdmissionSystem.Controllers.sub_classes.Admin_classes
                 FK_Department = FillSellectionDepartment(),
                 FK_type_Of_High_School_Cirtificate=FillSellectionType_of_high_school___()
                 
+            };
+            return dep;
+        }
+        Department_relation_type_view_model GetAllEDit( Department d,Type_of_high_school_Cirtificate t)
+        {
+            var dep = new Department_relation_type_view_model
+            {
+                FK_Department = FillSellectionDepartmentEDit(d),
+                FK_type_Of_High_School_Cirtificate = FillSellectionType_of_high_school___EDIt(t)
+
             };
             return dep;
         }
